experiments/preproc-electron/python-server/pyproject.toml
[project]
name = "preproc-server"
version = "0.1.0"
description = "RapidOCR preprocessing/detection server for Electron testbed"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.1",
  "uvicorn[standard]>=0.35.0",
  "python-multipart>=0.0.20",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.1",
  "numpy>=2.2.0",
  "Pillow>=11.3.0",
  "opencv-python-headless>=4.12.0.88",
  "rapidocr-onnxruntime>=1.4.4"
]

[project.scripts]
preproc-server = "preproc_server.cli:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"
